requirements.txt
streamlit 
openai
requests
python-pptx
